#Modularize components and reorganize directory structure
- Grouped task-related components into a Task directory (Task.tsx, TaskModal.tsx, PriorityLabel.tsx).
- Created a Column directory for column-specific components.
- Moved KanbanBoard component into its own KanbanBoard directory.
- Added a common directory for reusable components (Button, InputField, SelectField).
- Organized non-UI logic into a logic directory for drag-and-drop-related logic.
- Updated file paths and imports across the project to reflect the new directory structure.
